+++ store/rootmulti/store_test.go
@@ -101,18 +101,15 @@ func TestCacheMultiStoreWithVersion(t *testing.T) {
 	require.Equal(t, kvStore.Get(k), v)
 
 	// add new module stores (store4 and store5) to multi stores and commit
-	key4, key5 := types.NewKVStoreKey("store4"), types.NewKVStoreKey("store5")
-	ms.MountStoreWithDB(key4, types.StoreTypeIAVL, nil)
-	ms.MountStoreWithDB(key5, types.StoreTypeIAVL, nil)
+	ms.MountStoreWithDB(types.NewKVStoreKey("store4"), types.StoreTypeIAVL, nil)
+	ms.MountStoreWithDB(types.NewKVStoreKey("store5"), types.StoreTypeIAVL, nil)
 	err = ms.LoadLatestVersionAndUpgrade(&types.StoreUpgrades{Added: []string{"store4", "store5"}})
 	require.NoError(t, err)
 	ms.Commit()
 
 	// cache multistore of version before adding store4 should works
-	cms2, err := ms.CacheMultiStoreWithVersion(1)
+	_, err = ms.CacheMultiStoreWithVersion(1)
 	require.NoError(t, err)
-
-	require.Empty(t, cms2.GetKVStore(key4).Get([]byte("key")))
 
 	// require we cannot commit (write) to a cache-versioned multi-store
 	require.Panics(t, func() {